Create class to encapsulate the code executed in Main
The ReportConsoleProgram class was created for the purpose mentioned
above. The Main method is responsible for creating the Parser and
passing it along with the arguments. The ReportConsoleProgram will then
create the initial trello objects required for the report to be
generated.

According to the arguments passed, the respective report will be
generated. As of now, there are 2 available reports:

1. List of cards of a given list, with name, estimate and id;
2. Card count of all Lists of a given Board;